+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
  - Improvements to `LPCode`:
     - **(fix)** `LPCode`'s single `repr` function did not to enforce CSS orthogonality for non-commutative group algebras, producing invalid stabilizer matrices in non-abelian group-based `LPCode` cases.
-    - **(breaking)** changed representation (`repr`) to require separate `A_repr` (right regular representation) and `B_repr` (left regular representation) arguments in all `LPCode` constructors.
+    - Added separate `A_repr` (right regular representation) and `B_repr` (left regular representation) arguments in all `LPCode` constructors.
 - **(fix)** The gates `SQRTY`, `CXYZ`, `CZYX` were computing phases incorrectly when acting on `I` stabilizers.
 - **(fix)** Paulis with imaginary phases had their phases incorrectly tracked.
 - **(fix)** `rowdecompose` was not accounting for the phase of the input Pauli string, leading to potential errors in non-Clifford functionality.